[docker] Disable docker core check on macos
