+++ pkg/collector/corechecks/containers/docker_rate_capping_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2020 Datadog, Inc.
 
-// +build docker
+// +build docker,!darwin
 
 package containers
